chore: Move getJsonUserByName into signupSlice. Remove other dead code.
